Split main.ts code into files
